requirements.txt
alibabacloud_credentials
alibabacloud_sts20150401
alibabacloud_tea_console
fastapi[standard]~=0.115.12
markdown-it-py[plugins]~=3.0.0
openpyxl
pandas~=2.2.3
psycopg2-binary
pyjwt~=2.10.1
python-dotenv~=1.1.0
python-ulid~=3.0.0
sqlalchemy~=2.0.40
starlette~=0.46.1